Protected routes.

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,13 +1,22 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
 //date-fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if (!user) {
+      return
+    }
+
     const response = await fetch(`/api/workouts/${workout._id}`, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        Authorization: `Bearer ${user.token}`
+      }
     })
 
     // json is the item that we have deleted
